+++ apps/web/src/app/[lang]/(main)/(unirefund)/parties/refund-points/[partyId]/contracts/[contractId]/contract/_components/form.tsx
@@ -1,0 +1,221 @@
+"use client";
+
+import { toast } from "@/components/ui/sonner";
+import type {
+  UniRefund_ContractService_Refunds_RefundFeeHeaders_RefundFeeHeaderInformationDto as AssignableRefundFeeHeaders,
+  UniRefund_ContractService_ContractsForRefundPoint_ContractHeaders_ContractHeaderDetailForRefundPointDto as ContractHeaderDetailForRefundPointDto,
+  UniRefund_ContractService_ContractsForRefundPoint_ContractHeaders_ContractHeaderForRefundPointCreateDto as ContractHeaderForRefundPointCreateDto,
+} from "@ayasofyazilim/saas/ContractService";
+import { $UniRefund_ContractService_ContractsForRefundPoint_ContractHeaders_ContractHeaderForRefundPointCreateDto as $ContractHeaderForRefundPointCreateDto } from "@ayasofyazilim/saas/ContractService";
+import type { UniRefund_LocationService_AddressCommonDatas_AddressCommonDataDto as AddressTypeDto } from "@ayasofyazilim/saas/LocationService";
+import {
+  ActionButton,
+  ActionList,
+} from "@repo/ayasofyazilim-ui/molecules/action-button";
+import ConfirmDialog from "@repo/ayasofyazilim-ui/molecules/confirm-dialog";
+import { SchemaForm } from "@repo/ayasofyazilim-ui/organisms/schema-form";
+import { CustomComboboxWidget } from "@repo/ayasofyazilim-ui/organisms/schema-form/widgets";
+import {
+  handleDeleteResponse,
+  handlePostResponse,
+  handlePutResponse,
+} from "@repo/utils/api";
+import { CheckCircle, CircleX, Trash } from "lucide-react";
+import { useParams, useRouter } from "next/navigation";
+import type { TransitionStartFunction } from "react";
+import { useTransition } from "react";
+import type { ContractServiceResource } from "@/language-data/unirefund/ContractService";
+import { putRefundPointContractHeadersByIdMakePassiveApis } from "@/actions/unirefund/ContractService/put-actions";
+import {
+  postRefundPointContractHeadersById,
+  postRefundPointContractHeaderValidateByHeaderId,
+} from "@/actions/unirefund/ContractService/post-actions";
+import { deleteRefundPointContractHeadersById } from "@/actions/unirefund/ContractService/delete-actions";
+import { RefundFeeHeadersField } from "../../../_components/refund-fee-headers-field";
+
+export default function RefundPointContractHeaderUpdateForm({
+  addressList,
+  languageData,
+  refundFeeHeaders,
+  contractHeaderDetails,
+}: {
+  addressList: AddressTypeDto[];
+  languageData: ContractServiceResource;
+  refundFeeHeaders: AssignableRefundFeeHeaders[];
+  contractHeaderDetails: ContractHeaderDetailForRefundPointDto;
+}) {
+  const router = useRouter();
+  const { partyId } = useParams<{
+    partyId: string;
+  }>();
+  const [isPending, startTransition] = useTransition();
+  const uiSchema = {
+    "ui:className": "md:grid md:gap-2 md:grid-cols-2",
+    webSite: {
+      "ui:className": "md:col-span-full",
+      "ui:options": {
+        inputType: "url",
+      },
+    },
+    addressCommonDataId: {
+      "ui:className": "row-start-2",
+      "ui:widget": "address",
+    },
+    status: {
+      "ui:className": "md:col-span-full",
+    },
+    earlyRefund: {
+      "ui:widget": "switch",
+    },
+    refundFeeHeaders: {
+      "ui:className": "md:col-span-full",
+      "ui:field": "RefundFeeHeadersField",
+    },
+  };
+  return (
+    <div className="space-y-2">
+      <ContractActions
+        contractDetails={contractHeaderDetails}
+        isPending={isPending}
+        languageData={languageData}
+        startTransition={startTransition}
+      />
+      <SchemaForm<ContractHeaderForRefundPointCreateDto>
+        disabled={isPending}
+        fields={{
+          RefundFeeHeadersField: RefundFeeHeadersField(
+            refundFeeHeaders,
+            contractHeaderDetails.refundFeeHeaders.map((x) => ({
+              refundFeeHeaderId: x.id,
+              ...x,
+            })),
+          ),
+        }}
+        formData={{
+          ...contractHeaderDetails,
+          refundFeeHeaders: contractHeaderDetails.refundFeeHeaders.map((x) => ({
+            refundFeeHeaderId: x.id,
+            ...x,
+          })),
+        }}
+        onSubmit={({ formData: editedFormData }) => {
+          if (!editedFormData) return;
+          startTransition(() => {
+            void postRefundPointContractHeadersById({
+              id: partyId,
+              requestBody: editedFormData,
+            }).then((response) => {
+              handlePostResponse(response, router, {
+                prefix: `/parties/refund-points/${partyId}/contracts`,
+                suffix: "contract",
+                identifier: "id",
+              });
+            });
+          });
+        }}
+        schema={$ContractHeaderForRefundPointCreateDto}
+        uiSchema={uiSchema}
+        widgets={{
+          address: CustomComboboxWidget<AddressTypeDto>({
+            list: addressList,
+            languageData,
+            selectIdentifier: "id",
+            selectLabel: "fullAddress",
+          }),
+        }}
+      />
+    </div>
+  );
+}
+
+function ContractActions({
+  contractDetails,
+  languageData,
+  isPending,
+  startTransition,
+}: {
+  contractDetails: ContractHeaderDetailForRefundPointDto;
+  languageData: ContractServiceResource;
+  isPending: boolean;
+  startTransition: TransitionStartFunction;
+}) {
+  const router = useRouter();
+  return (
+    <ActionList>
+      <ActionButton
+        icon={CheckCircle}
+        loading={isPending}
+        onClick={() => {
+          startTransition(() => {
+            void postRefundPointContractHeaderValidateByHeaderId(
+              contractDetails.id,
+            ).then((response) => {
+              if (response.type === "success" && response.data) {
+                toast.success(
+                  languageData["Contracts.Actions.Validate.Success"],
+                );
+              } else {
+                toast.error(response.message);
+              }
+            });
+          });
+        }}
+        text={languageData["Contracts.Actions.Validate"]}
+      />
+      {!contractDetails.isDraft && !contractDetails.isActive && (
+        <ConfirmDialog
+          confirmProps={{
+            variant: "destructive",
+            children: languageData["Contracts.Actions.MakePassive"],
+            closeAfterConfirm: true,
+            onConfirm: () => {
+              startTransition(() => {
+                void putRefundPointContractHeadersByIdMakePassiveApis(
+                  contractDetails.id,
+                ).then((response) => {
+                  handlePutResponse(response, router, "../../");
+                });
+              });
+            },
+          }}
+          description={
+            languageData["Contracts.Actions.MakePassive.Description"]
+          }
+          title={languageData["Contracts.Actions.MakePassive.Title"]}
+          type="without-trigger"
+        >
+          <ActionButton
+            icon={CircleX}
+            loading={isPending}
+            text={languageData["Contracts.Actions.MakePassive"]}
+          />
+        </ConfirmDialog>
+      )}
+      <ConfirmDialog
+        confirmProps={{
+          variant: "destructive",
+          children: languageData["Contracts.Actions.Delete"],
+          closeAfterConfirm: true,
+          onConfirm: () => {
+            startTransition(() => {
+              void deleteRefundPointContractHeadersById(
+                contractDetails.id,
+              ).then((response) => {
+                handleDeleteResponse(response, router, "../../");
+              });
+            });
+          },
+        }}
+        description={languageData["Contracts.Actions.Delete.Description"]}
+        title={languageData["Contracts.Actions.Delete.Title"]}
+        type="without-trigger"
+      >
+        <ActionButton
+          icon={Trash}
+          loading={isPending}
+          text={languageData["Contracts.Actions.Delete"]}
+        />
+      </ConfirmDialog>
+    </ActionList>
+  );
+}
